:construction: Correct implementation of TRPO, Config Files for experiments and fixes
We included the clean implementation of the TRPO algorithm using the 3rd party dependency we will need to cite (torchtrustncg) with some slight modifications. We also have a model that groups features in related groups to create a feature extraction with a prior grouping of related observations. Some fiex throughout the code.