:wrench: Added configuration files and fixed things in the code
Create config files to run wach set of experiments through sweeps and a script to launch them. The advantage is that the sweep can have track of all the experiments (seeds) in separate process through the agents. Fixed a couple bugs in algo files and the logger. Making sure it backups the relevant files in wdb as well.